requirements.txt
pyyaml==6.0.1
fastapi
uvicorn
unidecode
numpy
torch>=1.7.0,<2.0.1,!=1.12.0
torchvision
pandas
requests
Pillow==9.5.0
opencv-python==4.5.5.64
tqdm
matplotlib
seaborn
scipy
gdown
einops
python-multipart
fuzzywuzzy
python-Levenshtein